backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "clerk-backend-api>=5.0.2",
    "fastapi>=0.125.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
